Added several colours.
Generation of boxes with different colours by BoxManager.

Player can switch between the coloured boxes using 1, 2, 3 keys.

Began implementing a box to depict the player's current colour choice: there is a problem that once you set a texture to one already set, it seems to re-allocate to that object by reference and so the
object it was taken from, actually no longer has it at the end of the application running. This problem must be fixed for the game to run. Currently it will not run in this state.